Make preprocessor submodules private
