+++ esmvaltool/preprocessor/_mask.py
