+++ tests/integration/preprocessor/_regrid/__init__.py
@@ -1,0 +1,4 @@
+"""
+Integration tests for the :mod:`esmvaltool.preprocessor._regrid` module.
+
+"""
