+++ tests/integration/preprocessor/_regrid/test_vinterp.py
@@ -12,8 +12,8 @@ import iris
 import numpy as np
 
 import tests
-from esmvaltool.preprocessor.regrid import _MDI, vinterp
-from tests.unit.preprocessor.regrid import _make_cube, _make_vcoord
+from esmvaltool.preprocessor._regrid import _MDI, vinterp
+from tests.unit.preprocessor._regrid import _make_cube, _make_vcoord
 
 
 class Test(tests.Test):
