--- tests/integration/preprocessor/regrid/__init__.py
@@ -1,4 +1,0 @@
-"""
-Integration tests for the :mod:`esmvaltool.preprocessor.regrid` module.
-
-"""
