+++ tests/unit/preprocessor/_derive/test_p_level_widths.py
@@ -3,7 +3,7 @@ from __future__ import division, print_function
 import numpy as np
 import pytest
 
-from esmvaltool.preprocessor.derive import _p_level_widths
+from esmvaltool.preprocessor._derive import _p_level_widths
 
 
 def test_col_is_not_monotonic():
